support java dependencies
