--- test/_functions/common/json-parser-with-function-config/java/.gitignore
@@ -1,2 +1,0 @@
-.gradle
-build
