+++ test/_functions/common/json-parser-with-function-config/java/function.yaml
@@ -16,8 +16,9 @@ metadata:
   name: parser
 spec:
   runtime: java
-  handler: CurlInfo
+  handler: JsonParser
   build:
-    commands:
-      - apt-get update
-      - apt-get install -y curl
+    dependencies:
+    - "group: com.fasterxml.jackson.core, name: jackson-databind, version: 2.9.0"
+    - "group: com.fasterxml.jackson.core, name: jackson-core, version: 2.9.0"
+    - "group: com.fasterxml.jackson.core, name: jackson-annotations, version: 2.9.0"
