--- test/_functions/common/json-parser-with-function-config/java/src/main/java/CurlInfo.java
@@ -1,35 +1,0 @@
-import com.fasterxml.jackson.core.type.TypeReference;
-import com.fasterxml.jackson.databind.ObjectMapper;
-import io.nuclio.Context;
-import io.nuclio.Event;
-import io.nuclio.EventHandler;
-import io.nuclio.Response;
-
-import java.util.HashMap;
-import java.util.Map;
-
-
-public class CurlInfo implements EventHandler {
-    @Override
-    public Response handleEvent(Context context, Event event) {
-        try {
-            if (!curlInstalled()) {
-                return new Response().setBody("curl not installed").setStatusCode(500);
-            }
-
-            ObjectMapper mapper = new ObjectMapper();
-            TypeReference<HashMap<String, String>> typeRef = new TypeReference<HashMap<String, String>>() {
-            };
-
-            Map<String, String> request = mapper.readValue(event.getBody(), typeRef);
-            return new Response().setBody(request.get("return_this"));
-        } catch (Throwable err) {
-            return new Response().setBody(err.toString()).setStatusCode(500);
-        }
-    }
-
-    private boolean curlInstalled() throws Throwable {
-        Process process = new ProcessBuilder("curl", "--version").start();
-        return process.waitFor() == 0;
-    }
-}
