--- test/_functions/common/reverser/java/.gitignore
@@ -1,2 +1,0 @@
-.gradle
-build
