+++ test/_functions/common/reverser/java/Reverser.java
@@ -3,7 +3,6 @@ import io.nuclio.Event;
 import io.nuclio.EventHandler;
 import io.nuclio.Response;
 
-
 public class Reverser implements EventHandler {
     @Override
     public Response handleEvent(Context context, Event event) {
