--- test/_functions/common/reverser/java/build.gradle
@@ -1,27 +1,0 @@
-plugins {
-  id 'com.github.johnrengelman.shadow' version '2.0.2'
-  id 'java'
-}
-
-repositories {
-    mavenCentral()
-}
-
-dependencies {
-    compile group: 'com.google.code.gson', name: 'gson', version: '2.8.2'
-    compile group: 'commons-cli', name: 'commons-cli', version: '1.4'
-
-    compile files('./nuclio-sdk-1.0-SNAPSHOT.jar')
-}
-
-// Create Jar in root directory
-tasks.withType(Jar) {
-    destinationDir = file("$rootDir")
-}
-
-shadowJar {
-   baseName = 'reverser'
-   classifier = null  // Don't append "all" to jar name
-}
-
-task nuclioJar(dependsOn: shadowJar)
